Added (commented out) chef section for checking out the app files
this is working on getting the deploy key stuff so OpsWorks can checkout
the git repo from GitHub so we can put the app files in the correct
places.  I'm still not convinced this is the right thing to do, but it
seems to be at least one thing to do, so I'm starting here.