using local files for the ssh/git wrapper and deploy key, checkout repo
local files won't work with AWS OpWorks, but one step at a time; this
does successfully checkout a git report from bitbucket; bitbucket allows
read-only deploy keys (unlike GitHub), so that was a safer test.

This relies on the deploy private key and the ssh wrapper; I'll check in
the ssh wrapper, but not the deploy key.